thinking machine/requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
anthropic==0.7.8
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6

# Frontend dependencies
streamlit==1.28.1
requests==2.31.0

